+++ packages/luthor/lib/src/annotations/validators/custom_with_schema.dart
@@ -1,0 +1,6 @@
+class WithSchemaCustomValidator {
+  final String? message;
+  final bool Function(Object? value, Map<String, Object?> data) customValidator;
+
+  const WithSchemaCustomValidator(this.customValidator, {this.message});
+}